+++ src/sklearn_evaluation/plot/__init__.py
@@ -11,11 +11,15 @@ from .grid_search import grid_search
 from .validation_curve import validation_curve
 from .learning_curve import learning_curve
 from .metrics import metrics_at_thresholds
+from .clustering import elbow_curve, elbow_curve_from_results, silhouette_analysis, silhouette_analysis_from_results
 from .regression import residuals, prediction_error
 from .clustering import elbow_curve, elbow_curve_from_results
 from .feature_ranking import Rank1D, Rank2D
+from .target_analysis import target_analysis
 from .calibration import calibration_curve, scores_distribution
 from .classification_report import classification_report, ClassificationReport
+from .ks_statistics import ks_statistic
+from .cumulative_gain_lift_curve import cumulative_gain, lift_curve
 
 __all__ = [
     'confusion_matrix',
@@ -32,10 +36,16 @@ __all__ = [
     'prediction_error',
     'elbow_curve',
     'elbow_curve_from_results',
+    'silhouette_analysis',
+    'silhouette_analysis_from_results',
     'calibration_curve',
+    'target_analysis',
     'scores_distribution',
     'classification_report',
     'ClassificationReport',
     'Rank1D',
     'Rank2D'
+    'ks_statistic',
+    'cumulative_gain',
+    'lift_curve'
 ]