+++ tests/conftest.py
@@ -18,6 +18,7 @@ from matplotlib.testing.conftest import (  # noqa
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.model_selection import GridSearchCV
 from sklearn import datasets
+from sklearn.model_selection import train_test_split
 
 
 def _path_to_tests():
@@ -123,3 +124,39 @@ def grid_search_param_with_none():
     clf.fit(X, y)
 
     return clf
+
+
+@pytest.fixture(scope='session')
+def target_analysis_binary():
+    kwargs = {
+        "n_samples": 100,
+        "n_classes": 2,
+        "n_features": 20,
+        "n_informative": 8,
+        "n_redundant": 2,
+        "n_clusters_per_class": 1,
+        "random_state": 89092,
+    }
+
+    X, y = datasets.make_classification(**kwargs)
+
+    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=101)
+    return X_train, X_test, y_train, y_test
+
+
+@pytest.fixture(scope='session')
+def target_analysis_multiclass():
+    kwargs = {
+        "n_samples": 100,
+        "n_classes": 5,
+        "n_features": 20,
+        "n_informative": 8,
+        "n_redundant": 2,
+        "n_clusters_per_class": 1,
+        "random_state": 89092,
+    }
+
+    X, y = datasets.make_classification(**kwargs)
+
+    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=101)
+    return X_train, X_test, y_train, y_test